[REFACTOR] Refactoring constant
